refactor(plugin): split plugin.rs into plugin/ module
Mechanically extract the 3,296-line src/plugin.rs into src/plugin/mod.rs
plus 10 submodules (install, update, remove, list, search, run_plugin,
create, validate, publish, tests). All public API paths are preserved:
PluginAction, PluginOptions, PluginCapabilities, run(), resolve_plugin_command(),
and run_lifecycle_hook() remain importable from crate::plugin::.

Update specs/plugin/plugin.spec.md to reference src/plugin/mod.rs.
